+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## [Unreleased]
 ### Added
  - Added SPlusGlobalEvents framework, for initializing from s+ and future s+ eventing.
+ - Added OnSettingsApplied and OnSettingsCleared events to AbstractSPlusOriginatorShim
 
 ## [5.0.0] - 2018-05-09
 ### Changed